Backend - Commit programming

+++ AutoLend.Application/Services/Car/CarService.cs
@@ -1,23 +1,35 @@
 ﻿using AutoLend.Core.ApiModels.Car;
+using AutoLend.Core.Esceptions;
 using AutoLend.Data.CoreModels.Car;
 using AutoLend.Data.DataModels.Car;
 using AutoLend.Data.Repositories.Car;
+using AutoLend.Data.Repositories.Model;
 using Newtonsoft.Json;
-using System.Text.Json.Serialization;
-using JsonConverter = Newtonsoft.Json.JsonConverter;
+using System.Data;
 
 namespace AutoLend.Core.Services.Car {
     internal class CarService : ICarService {
 
         private readonly ICarRepository _carRepository;
-
-        public CarService( ICarRepository carRepository ) {
+        private readonly IModelRepository _modelRepository;
+        public CarService( ICarRepository carRepository, IModelRepository modelRepository ) {
             _carRepository = carRepository;
+            _modelRepository = modelRepository;
         }
 
         public async Task CreateCar( CarCreateRequest car ) {
+
+            if (await _carRepository.LicensePlateExistsAsync(car.LicensePlate))
+                throw new BusinessException("License plate already exist.");
+
+            var ModelId = (await _modelRepository.GetByModelNameAsync(car.ModelName)) ?? throw new BusinessException("Model name not found");
+            
+            if (car.Year > DateTime.Now.Year)
+                if (car.Year > DateTime.Now.Year)
+                    throw new BusinessException("Wrong year.");
+
             CarCreateDTO CarDto = new() {
-                ModelName = car.ModelName,
+                ModelId = ModelId.Id,
                 Year = car.Year,
                 LicensePlate = car.LicensePlate,
                 IsAvailable = car.IsAvailable,
@@ -27,6 +39,10 @@ namespace AutoLend.Core.Services.Car {
         }
 
         public async Task DeleteCar( int carId ) {
+
+            if (await _carRepository.GetByIdAsync(carId) is null)
+                throw new BusinessException("Car not found or is not active.");
+
             await _carRepository.DeleteAsync(carId);
         }
 
@@ -35,16 +51,16 @@ namespace AutoLend.Core.Services.Car {
         }
 
         public async Task<Data.DataModels.Car.Car?> GetCarById( int carId ) {
-            return await _carRepository.GetByIdAsync(carId);
+            return await _carRepository.GetByIdAsync(carId) ?? throw new BusinessException("Car not found");
         }
 
         public async Task<IEnumerable<CarSearch?>> Search( CarSearchRequest car ) {
 
-            var modelIds = car.ModelIds?.Select(x => new ModelIdItem() { 
+            var modelIds = car.ModelIds?.Select(x => new ModelIdItem() {
                 ModelId = x,
             }).ToList();
 
-            var modelIdsJson = modelIds is null ? string.Empty: JsonConvert.SerializeObject(modelIds);
+            var modelIdsJson = modelIds is null ? string.Empty : JsonConvert.SerializeObject(modelIds);
 
             CarSearchDTO CarDto = new() {
                 ModelIdsJSON = modelIdsJson,
@@ -62,12 +78,32 @@ namespace AutoLend.Core.Services.Car {
         }
 
         public async Task UpdateCar( int carId, CarUpdateRequest car ) {
+
+
+            var carObject = await _carRepository.GetByIdAsync(carId);
+
+            if (carObject is null)
+                throw new BusinessException("Car not found.");
+
+            if (!string.IsNullOrEmpty(car.LicensePlate))
+                if (await _carRepository.LicensePlateExistsAsync(car.LicensePlate, carId))
+                    throw new BusinessException("License plate already exist.");
+
+            if (!string.IsNullOrEmpty(car.ModelName))
+                if (await _modelRepository.GetByModelNameAsync(car.ModelName) is null)
+                    throw new BusinessException("Model name not found.");
+
+            if (car.Year.HasValue && car.Year > DateTime.Now.Year)
+                if (car.Year > DateTime.Now.Year)
+                    throw new BusinessException("Wrong year.");
+
             CarUpdateDTO CarDto = new() {
                 Id = carId,
                 ModelName = car.ModelName,
                 Year = car.Year,
                 LicensePlate = car.LicensePlate,
                 IsAvailable = car.IsAvailable,
+                Cost = car.Cost
             };
 
             await _carRepository.UpdateAsync(CarDto);
